npmignore
Required to publish compiled javascript code.

Without .npmignore, npm uses the .gitignore files, with the result
that all files excluded from git are also excluded from the package
published to npm. This does not work for this project, because the
compiled javascript files under /lib are excluded from version control,
but need to be included in the published package.